feat(105): fix error on history folder
chore(105)made sections of dashboard-overview page into components for better code readability
