+++ scrybe_frontend/src/pages/History/assets/images/index.jsx
@@ -1,7 +1,7 @@
 export { default as searchIcon } from "./search-icon.svg";
 export { default as profileUpload } from "./Profile-upload.svg";
 export { default as uploadIcon } from "./Upload-icon.svg";
-export { default as gridIcon } from "./icongrid.png";
+export { default as gridIcon } from "./Icongrid.png";
 export { default as analysisIcon } from "./analysis.svg";
 export { default as buggerMenu } from "./iconbugger.png";
 export { default as closeIcon } from "./iconclose.png";
